refactor(mappers): reorganize mappers by layers
