+++ app/src/main/java/com/example/shmr_finance_app_android/data/remote/mapper/AccountRemoteMapper.kt
@@ -1,0 +1,33 @@
+package com.example.shmr_finance_app_android.data.remote.mapper
+
+import com.example.shmr_finance_app_android.data.model.AccountDTO
+import com.example.shmr_finance_app_android.data.model.StatItemDTO
+import com.example.shmr_finance_app_android.data.remote.model.AccountResponse
+import com.example.shmr_finance_app_android.data.remote.model.StatItemResponse
+import dagger.Reusable
+import javax.inject.Inject
+
+@Reusable
+class AccountRemoteMapper @Inject constructor() {
+    fun mapAccount(response: AccountResponse): AccountDTO {
+        return AccountDTO(
+            id = response.id,
+            name = response.name,
+            balance = response.balance,
+            currency = response.currency,
+            incomeStats = response.incomeStats.map { mapStatItem(it) },
+            expenseStats = response.expenseStats.map { mapStatItem(it) },
+            createdAt = response.createdAt,
+            updatedAt = response.updatedAt
+        )
+    }
+
+    private fun mapStatItem(item: StatItemResponse): StatItemDTO {
+        return StatItemDTO(
+            categoryId = item.categoryId,
+            categoryName = item.categoryName,
+            emoji = item.emoji,
+            amount = item.amount
+        )
+    }
+}
